requirements.txt
billiard==3.6.4.0
biopython==1.79
bokeh==2.3.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
evcouplings==0.1.1
filelock==3.4.1
idna==3.3
igraph==0.9.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.1
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mmtf-python==1.1.2
msgpack==1.0.3
numba==0.53.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
pkg_resources==0.0.0
propy3==1.1.0
psutil==5.9.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
sklearn==0.0
texttable==1.6.4
threadpoolctl==3.1.0
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
zipp==3.6.0